tests: refactor tests into separate files
